Create reference routes
Set up reference routes and a Reference controller as a place to put
actions that are only used as a reference for functionality later down
the line. Moved all of the mongo query code to this new controller.
