+++ conf/routes
@@ -2,6 +2,15 @@
 # This file defines all application routes (Higher priority routes first)
 # ~~~~
 
+#==================
+# REFERENCE ROUTES
+#==================
+# Database query reference
+GET     /ref/dbquery                controllers.Reference.dbQuery
+
+#================
+# WHEREU@ ROUTES
+#================
 # Account verification
 POST    /account/request            controllers.Whereuat.requestAccount
